+++ examples/webpack-bundler/webpack.config.js
@@ -8,7 +8,5 @@ module.exports = {
     filename: "bootstrap.js",
   },
   mode: "development",
-  plugins: [
-    new CopyWebpackPlugin(['index.html'])
-  ],
+  plugins: [new CopyWebpackPlugin(["index.html"])],
 };