+++ src/kits/revocation.ts
@@ -1,21 +1,9 @@
-import { Signer } from 'umbral-pre';
-
-import { RevocationOrder, TreasureMap } from '../policies/collections';
-import { ChecksumAddress } from '../types';
+import { RevocationOrder, Signer, TreasureMap } from '@nucypher/nucypher-core';
 
 export class RevocationKit {
-  public revocations: Record<ChecksumAddress, RevocationOrder>;
+  public revocationOrders: RevocationOrder[];
 
   constructor(treasureMap: TreasureMap, signer: Signer) {
-    this.revocations = {};
-    Object.entries(treasureMap.destinations).forEach(
-      ([nodeId, encryptedKFrag]) => {
-        this.revocations[nodeId] = new RevocationOrder(
-          nodeId,
-          encryptedKFrag,
-          signer
-        );
-      }
-    );
+    this.revocationOrders = treasureMap.makeRevocationOrders(signer);
   }
 }